requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.2.1
email-validator==2.1.1
fastapi==0.88.0
h11==0.14.0
idna==3.4
motor==3.1.1
passlib[bcrypt]==1.7.4
pymongo==4.3.3
python-dotenv==0.21.0
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.12.2
uvicorn==0.20.0
